address most of the requested changes
More to come tomorrow
